requirements.txt

gradio==4.15.0
pytesseract
pillow
playwright
pandas
beautifulsoup4
openai
